refactor: give every command and extension a file of its own
The package puts the handler of a command in the file of that command:
handleDATA in data.go, handleAUTH in auth.go, handleXCLIENT in
xclient.go, handleBDAT in chunk.go. Two handlers stood outside that
order, and the test files of the package name the files that they wanted:
starttls_test.go, verify_session_test.go, reply_test.go, smtputf8_test.go.

handleSTARTTLS moves to starttls.go, and handleVRFY joins the
Verification type in verify.go.

reply.go takes the writers of a reply from session.go, with the reader of
a status code and the guard that takes a control character out of the
text. session.go keeps the session: the reads, the serve loop, and the
end of a session.

smtputf8.go takes what RFC 6531 asks of an address. The two handlers
carried that reading inline, and they now call one function each, so the
policy of the extension stands in the file of the extension.

extensions() moves next to handleEHLO, which is the one caller.

The replies to an address that does not read move to address.go, next to
the reader that answers with them. The two handlers wrote the same reply
in words of their own.

Nothing changes on the wire. Every declaration of the package is the same
one in another file, but for the two functions that hold the reading of a
character set and the two replies that they name.
